+++ com.stansassets.scene-management/Runtime/Statis/AdditiveScenesLoader.cs
@@ -385,40 +385,9 @@ namespace StansAssets.SceneManagement
         {
             return BuildConfigurationSettings.Instance.Configuration.IsSceneAddressable(sceneName);
         }
-
-        class AddressableSceneUnloader
-        {
-            readonly SceneInstance m_SceneInstance;
-            readonly string m_SceneName;
-
-            Action<AddressableSceneUnloader> m_Complete;
-            AsyncOperationHandle<SceneInstance> m_AsyncOperationHandle;
-
-            public AddressableSceneUnloader(SceneInstance sceneInstance)
-            {
-                m_SceneInstance = sceneInstance;
-                m_SceneName = string.Copy(m_SceneInstance.Scene.name);
-            }
-
-            public void Unload(Action<AddressableSceneUnloader> complete)
-            {
-                m_Complete = complete;
-                var asyncOperationHandle = Addressables.UnloadSceneAsync(m_SceneInstance);
-                asyncOperationHandle.Completed += AddressableSceneUnloaded;
-            }
-
-            void AddressableSceneUnloaded(AsyncOperationHandle<SceneInstance> asyncOperation)
-            {
-                m_AsyncOperationHandle = asyncOperation;
-                m_Complete.Invoke(this);
-            }
-
-            public string SceneName => m_SceneName;
-            public AsyncOperationHandle<SceneInstance> AsyncOperationHandle => m_AsyncOperationHandle;
-        }
     }
 
-    public static class AddressablesLogger
+    static class AddressablesLogger
     {
         public static bool Verbose = false;
